requirements.txt
requests~=2.32.3
pytz~=2024.2
python-dotenv~=1.0.1